Refactor discord container into a mixin
